Split connection test helpers
